[Storage Service] Move utility functions into utils.rs
