+++ state-sync/storage-service/server/src/lib.rs
@@ -41,6 +41,7 @@ mod moderator;
 pub mod network;
 mod optimistic_fetch;
 pub mod storage;
+mod utils;
 
 #[cfg(test)]
 mod tests;
