+++ state-sync/storage-service/server/src/tests/mock.rs
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    metrics, network::StorageServiceNetworkEvents, storage::StorageReader, StorageServiceServer,
+    metrics, network::StorageServiceNetworkEvents, storage::StorageReader, tests::utils,
+    StorageServiceServer,
 };
 use anyhow::Result;
 use aptos_channels::{aptos_channel, message_queues::QueueStyle};
 use aptos_config::{config::StorageServiceConfig, network_id::NetworkId};
 use aptos_crypto::HashValue;
-use aptos_logger::Level;
 use aptos_network::{
     application::{interface::NetworkServiceEvents, storage::PeersAndMetadata},
     peer_manager::PeerManagerNotification,
@@ -70,7 +70,7 @@ impl MockClient {
         MockTimeService,
         Arc<PeersAndMetadata>,
     ) {
-        initialize_logger();
+        utils::initialize_logger();
 
         // Create the storage reader
         let storage_config = storage_config.unwrap_or_default();
@@ -211,14 +211,6 @@ fn get_random_network_id() -> NetworkId {
         2 => NetworkId::Public,
         num => panic!("This shouldn't be possible! Got num: {:?}", num),
     }
-}
-
-/// Initializes the Aptos logger for tests
-fn initialize_logger() {
-    aptos_logger::Logger::builder()
-        .is_async(false)
-        .level(Level::Debug)
-        .build();
 }
 
 // This automatically creates a MockDatabaseReader.
