+++ state-sync/storage-service/server/src/tests/new_transactions.rs
@@ -6,14 +6,10 @@ use aptos_config::{
     config::StorageServiceConfig,
     network_id::{NetworkId, PeerNetworkId},
 };
-use aptos_storage_service_types::{
-    requests::{DataRequest, NewTransactionsWithProofRequest, StorageServiceRequest},
-    responses::DataResponse,
+use aptos_storage_service_types::requests::{
+    DataRequest, NewTransactionsWithProofRequest, StorageServiceRequest,
 };
-use aptos_types::{
-    epoch_change::EpochChangeProof, ledger_info::LedgerInfoWithSignatures,
-    transaction::TransactionListWithProof, PeerId,
-};
+use aptos_types::{epoch_change::EpochChangeProof, PeerId};
 use claims::assert_none;
 use futures::channel::oneshot::Receiver;
 
@@ -82,7 +78,7 @@ async fn test_get_new_transactions() {
             .await;
 
             // Verify a response is received and that it contains the correct data
-            verify_new_transactions_with_proof(
+            utils::verify_new_transactions_with_proof(
                 &mut mock_client,
                 response_receiver,
                 transaction_list_with_proof,
@@ -188,14 +184,14 @@ async fn test_get_new_transactions_different_networks() {
             .await;
 
             // Verify a response is received and that it contains the correct data for both peers
-            verify_new_transactions_with_proof(
+            utils::verify_new_transactions_with_proof(
                 &mut mock_client,
                 response_receiver_1,
                 transaction_list_with_proof_1,
                 highest_ledger_info.clone(),
             )
             .await;
-            verify_new_transactions_with_proof(
+            utils::verify_new_transactions_with_proof(
                 &mut mock_client,
                 response_receiver_2,
                 transaction_list_with_proof_2,
@@ -278,7 +274,7 @@ async fn test_get_new_transactions_epoch_change() {
         .await;
 
         // Verify a response is received and that it contains the correct data
-        verify_new_transactions_with_proof(
+        utils::verify_new_transactions_with_proof(
             &mut mock_client,
             response_receiver,
             transaction_list_with_proof,
@@ -347,7 +343,7 @@ async fn test_get_new_transactions_max_chunk() {
         .await;
 
         // Verify a response is received and that it contains the correct data
-        verify_new_transactions_with_proof(
+        utils::verify_new_transactions_with_proof(
             &mut mock_client,
             response_receiver,
             transaction_list_with_proof,
@@ -395,25 +391,4 @@ async fn get_new_transactions_with_proof_for_peer(
     mock_client
         .send_request(storage_request, peer_id, network_id)
         .await
-}
-
-/// Verifies that a new transactions with proof response is received
-/// and that the response contains the correct data.
-async fn verify_new_transactions_with_proof(
-    mock_client: &mut MockClient,
-    receiver: Receiver<Result<bytes::Bytes, aptos_network::protocols::network::RpcError>>,
-    expected_transactions_with_proof: TransactionListWithProof,
-    expected_ledger_info: LedgerInfoWithSignatures,
-) {
-    let storage_service_response = mock_client.wait_for_response(receiver).await.unwrap();
-    match storage_service_response.get_data_response().unwrap() {
-        DataResponse::NewTransactionsWithProof((transactions_with_proof, ledger_info)) => {
-            assert_eq!(transactions_with_proof, expected_transactions_with_proof);
-            assert_eq!(ledger_info, expected_ledger_info);
-        },
-        response => panic!(
-            "Expected new transaction with proof but got: {:?}",
-            response
-        ),
-    };
 }
